refactor(notifications): extract custom-sound capture into a hook
Move the recorder/stream/timer/preview lifecycle, decode + silence
detection, trim, and save out of AddCustomSoundDialog into a
useCustomSoundCapture hook, leaving the component as layout only. Group
the captured-clip controls (preview / trim / discard) into a Card.

No behaviour change; full @posthog/ui suite green (1052).

Generated-By: PostHog Code
Task-Id: 19173ab4-1301-4e3f-af36-5427dad0ada5
